refactor: extract WorkService and SubmissionService from PoolService
Phase 2 of issue #67. Introduces testable service classes that contain all
domain logic; PoolService becomes a thin HTTP adapter that parses requests,
holds the mutex, delegates, and serialises responses.

WorkService (include/puzzpool/work_service.hpp, src/work_service.cpp):
  - assignWork()  — full work assignment pipeline (upsert worker, reclaim
                    detection, existing chunk, test chunk, reclaim, allocate)
  - heartbeat()   — update last_seen + heartbeat_at
  - reclaimTimedOutChunks() — timeout sweep
  - upsertWorkerAndDetectReactivation(), existingAssignedChunk(),
    reclaimChunk(), claimTestChunk(), readChunk() — moved from PoolService

SubmissionService (include/puzzpool/submission_service.hpp, src/submission_service.cpp):
  - submitDone()  — underscan rejection, completion update
  - submitFound() — FOUND update, late-FOUND via reclaimed-path, deduplication,
                    BINGO file append
  - clearTestChunkIfNeeded() — moved from PoolService

PoolService changes:
  - service.hpp: removed private method declarations now owned by the services;
    added WorkService ws_ and SubmissionService ss_ members
  - service.cpp: initialises ws_(db_, allocator_) and ss_(db_) in ctor;
    reclaimTimedOutChunks() delegates to ws_
  - service_work.cpp: now only HTTP parsing + delegation (~50 loc)
  - service_submit.cpp: now only HTTP parsing + delegation (~30 loc)

Both service classes require only PoolDb& (and Allocator& for WorkService),
making them directly instantiable in unit tests with an in-memory SQLite DB.
