+++ include/puzzpool/service.hpp
@@ -3,18 +3,21 @@
 #include <puzzpool/allocator.hpp>
 #include <puzzpool/config.hpp>
 #include <puzzpool/db.hpp>
+#include <puzzpool/submission_service.hpp>
 #include <puzzpool/types.hpp>
+#include <puzzpool/work_service.hpp>
 
 #include <crow.h>
 #include <nlohmann/json.hpp>
 
-#include <cstdint>
 #include <mutex>
-#include <optional>
 #include <string>
 
 namespace puzzpool {
 
+// PoolService is the HTTP adapter layer. It parses requests, holds the mutex,
+// delegates business logic to WorkService / SubmissionService / Allocator,
+// and serialises responses. It owns no domain logic directly.
 class PoolService {
 public:
     explicit PoolService(const Config& cfg);
@@ -37,34 +40,21 @@ private:
     void ensureSingleActivePuzzle();
     void ensureAllocators();
 
-    bool upsertWorkerAndDetectReactivation(
-        const std::string& name, double hashrate, const std::string& version,
-        const std::optional<std::string>& minChunkKeys,
-        const std::optional<std::string>& chunkQuantumKeys);
-
-    std::optional<ChunkRow> existingAssignedChunk(const std::string& name, int64_t puzzleId);
-    std::optional<ChunkRow> reclaimChunk(const std::string& name, int64_t puzzleId);
-    std::optional<ChunkRow> claimTestChunk(const std::string& name, const PuzzleRow& puzzle);
-    ChunkRow                readChunk(SQLite::Statement& q);
-
-    crow::response handleSubmitDone(const std::string& name, int64_t jobId, const nlohmann::json& body);
-    crow::response handleSubmitFound(const std::string& name, int64_t jobId, const nlohmann::json& body);
-
-    void clearTestChunkIfNeeded(int64_t jobId);
-
     nlohmann::json buildStats(const PuzzleRow& puzzle);
     nlohmann::json puzzleJson(const PuzzleRow& p);
 
-    static std::string     formatDouble(double v);
-    static std::string     nowIsoUtc();
-    static crow::response  jsonResponse(const nlohmann::json& j, int code = 200);
-    static crow::response  errorJsonResponse(int code, const nlohmann::json& j);
-    static crow::response  errorResponse(int code, const std::string& msg);
+    static std::string    formatDouble(double v);
+    static std::string    nowIsoUtc();
+    static crow::response jsonResponse(const nlohmann::json& j, int code = 200);
+    static crow::response errorJsonResponse(int code, const nlohmann::json& j);
+    static crow::response errorResponse(int code, const std::string& msg);
 
-    const Config cfg_;
-    PoolDb       db_;
-    Allocator    allocator_;
-    std::mutex   mu_;
+    const Config       cfg_;
+    PoolDb             db_;
+    Allocator          allocator_;
+    WorkService        ws_;
+    SubmissionService  ss_;
+    std::mutex         mu_;
 };
 
 } // namespace puzzpool
